replace next 11 with next 12 version
