--- packages/starter/next-12-starter/package.json
@@ -1,27 +1,0 @@
-{
-    "name": "next-12-starter",
-    "private": true,
-    "scripts": {
-        "clean": "shx rm -rf .next",
-        "dev": "next dev",
-        "build": "yarn clean && next build",
-        "start": "next start",
-        "lint": "next lint"
-    },
-    "dependencies": {
-        "@solana/wallet-adapter-base": "^0.9.1",
-        "@solana/wallet-adapter-react": "^0.15.1",
-        "@solana/wallet-adapter-react-ui": "^0.9.1",
-        "@solana/wallet-adapter-wallets": "^0.14.1",
-        "next": "12.0.7",
-        "react": "17.0.2",
-        "react-dom": "17.0.2"
-    },
-    "devDependencies": {
-        "@types/node": "17.0.8",
-        "@types/react": "17.0.38",
-        "eslint": "8.6.0",
-        "eslint-config-next": "12.0.7",
-        "typescript": "4.5.4"
-    }
-}
