+++ packages/starter/nextjs-starter/next.config.js
@@ -1,8 +1,4 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const withTM = require('next-transpile-modules')(['@blocto/sdk', '@project-serum/sol-wallet-adapter']);
-
 /** @type {import('next').NextConfig} */
-module.exports = withTM({
-    reactStrictMode: true,
-    webpack5: true,
-});
+module.exports = {
+  reactStrictMode: true,
+}
