+++ packages/starter/nextjs-starter/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@solana/wallet-adapter-nextjs-starter",
-    "version": "0.7.1",
+    "version": "0.8.0",
     "author": "Solana Maintainers <maintainers@solana.foundation>",
     "repository": "https://github.com/solana-labs/wallet-adapter",
     "license": "Apache-2.0",
@@ -31,16 +31,15 @@
         "@solana/wallet-adapter-react": "^0.15.1",
         "@solana/wallet-adapter-react-ui": "^0.9.1",
         "@solana/wallet-adapter-wallets": "^0.14.1",
-        "@solana/web3.js": "^1.20.0",
-        "next": "^11.0.0",
-        "react": "^17.0.0",
-        "react-dom": "^17.0.0"
+        "next": "12.0.7",
+        "react": "17.0.2",
+        "react-dom": "17.0.2"
     },
     "devDependencies": {
-        "@types/react": "^17.0.0",
-        "eslint": "^7.32.0",
-        "eslint-config-next": "^11.0.0",
-        "next-transpile-modules": "^8.0.0",
+        "@types/node": "17.0.8",
+        "@types/react": "17.0.38",
+        "eslint": "8.6.0",
+        "eslint-config-next": "12.0.7",
         "shx": "^0.3.3",
         "typescript": "~4.4.4"
     }
