+++ packages/starter/nextjs-starter/pages/index.tsx
@@ -1,10 +1,10 @@
+import type { NextPage } from 'next';
 import { WalletDisconnectButton, WalletMultiButton } from '@solana/wallet-adapter-react-ui';
-import { NextPage } from 'next';
 import Head from 'next/head';
 import Image from 'next/image';
 import styles from '../styles/Home.module.css';
 
-const Index: NextPage = () => {
+const Home: NextPage = () => {
     return (
         <div className={styles.container}>
             <Head>
@@ -24,7 +24,7 @@ const Index: NextPage = () => {
                 </div>
 
                 <p className={styles.description}>
-                    Get started by editing <code className={styles.code}>pages/index.js</code>
+                    Get started by editing <code className={styles.code}>pages/index.tsx</code>
                 </p>
 
                 <div className={styles.grid}>
@@ -69,4 +69,4 @@ const Index: NextPage = () => {
     );
 };
 
-export default Index;
+export default Home;
